fead_product_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
whitenoise==6.6.0
django-redis==5.2.0
gunicorn==21.2.0
django-environ==0.10.0
python-dotenv
djangorestframework-simplejwt
selenium==4.18.0
webdriver-manager==4.0.1
requests>=2.25.0
python-telegram-bot>=20.0  # Optional, for more advanced features
Pillow>=10.0.0